avoid divide-by-zero
Division of two floats can be truncated to zero if cast as a long. finalNmberOfSteps is used as a denominator in subsequent calculations.
 This corrects issue #321.

+++ cnc_ctrl_v1/CNC_Functions.h
@@ -446,7 +446,7 @@ int   cordinatedMove(const float& xEnd, const float& yEnd, const float& zEnd, fl
     //compute feed details
     MMPerMin = constrain(MMPerMin, 1, MAXFEED);   //constrain the maximum feedrate, 35ipm = 900 mmpm
     float  stepSizeMM           = computeStepSize(MMPerMin);
-    long   finalNumberOfSteps   = abs(distanceToMoveInMM/stepSizeMM);
+    float   finalNumberOfSteps  = abs(distanceToMoveInMM/stepSizeMM);
     float  delayTime            = calculateDelay(stepSizeMM, MMPerMin);
     float  zFeedrate            = calculateFeedrate((zDistanceToMoveInMM/finalNumberOfSteps), delayTime);
